fix(pbzx)!: keep cpio filesize and mtime wider than 32 bits
odc stores both as 11 octal digits, a 33-bit range reaching 8 GiB, but
read_odc_header narrowed each with `as u32`. A file between 4 and 8 GiB
wrapped: 5_000_000_000 parsed as 705_032_704. Since filesize drives how
many bytes the reader consumes, a wrapped value desynchronises the stream
and corrupts every entry after it, and a wrapped mtime is a plausible but
wrong timestamp presented as fact.

libarchive parses the same field into a 64-bit value and says why:
"entry_bytes_remaining is at least 64 bits and therefore guaranteed to be
big enough for a 33-bit file size".

newc's fields are 8 hex digits, exactly 32 bits, so widening those is
lossless. The payload fixture stays well inside u32, so nothing observed
here was affected.

BREAKING CHANGE: CpioHeader::filesize, CpioHeader::mtime, FileEntry::mtime
and CpioEntry::mtime are now u64.

+++ dpp-python/src/types.rs
@@ -237,7 +237,7 @@ pub struct PyFileEntry {
     #[pyo3(get)]
     pub mode: u32,
     #[pyo3(get)]
-    pub mtime: u32,
+    pub mtime: u64,
     #[pyo3(get)]
     pub uid: u32,
     #[pyo3(get)]
